requirements.txt
Django==4.2.7
djangorestframework==3.14.0
transformers==4.35.2
torch==2.2.0
PyMuPDF==1.23.8
pdfminer.six==20221105
requests==2.31.0
python-dotenv==1.0.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
groq==0.4.1
Pillow==10.1.0
drf-spectacular==0.27.0
celery==5.3.4
redis==5.0.1
django-celery-beat==2.5.0
django-celery-results==2.5.0
django-filter==23.5
python-docx==0.8.11
python-pptx==0.6.21